Dont double log staff account creation + correctly export for hot refresh (#1003)
