+++ client/src/pages/lab_management/audit_logs/ManualRoomSignInModal.tsx
@@ -2,7 +2,7 @@ import { gql, useLazyQuery, useMutation, useQuery } from "@apollo/client";
 import PrettyModal from "../../../common/PrettyModal";
 import { Box, Button, CircularProgress, MenuItem, Select, Stack, TextField, Typography } from "@mui/material";
 import { useState } from "react";
-import { GET_USER_BY_USERNAME_OR_UID, PartialUser } from "../../../queries/getUsers";
+import { GET_USER_BY_USERNAME_OR_UID, PartialUser } from "../../../queries/userQueries";
 import AuditLogEntity from "../audit_logs/AuditLogEntity";
 import CloseIcon from '@mui/icons-material/Close';
 import CheckIcon from '@mui/icons-material/Check';
