+++ client/src/pages/lab_management/users/UsersPage.tsx
@@ -3,7 +3,7 @@ import SearchBar from "../../../common/SearchBar";
 import { Box, Button, Stack, Typography } from "@mui/material";
 import UserCard from "./UserCard";
 import { useLazyQuery, useQuery } from "@apollo/client";
-import { GET_NUM_USERS, GET_USERS_LIMIT, PartialUser } from "../../../queries/getUsers";
+import { GET_NUM_USERS, GET_USERS_LIMIT, PartialUser } from "../../../queries/userQueries";
 import RequestWrapper from "../../../common/RequestWrapper";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
